Add files via upload
We have a table in which types of SCDs and there respective SP names are given. We have to create one stored procedure (SP) with SCD_type as parameter and the SP will execute the respective SP mentioned in the table based on the parameter. If no parameter is provided then all SPs in the table should get executed.